Improve Jobs collection
Improved performance by querying system tables instead of SMO.
Scripting of jobs can be disabled
