+++ DBADash/SchemaSnapshotDB.cs
@@ -83,195 +83,15 @@ namespace DBADash
 
     }
 
-    public class SchemaSnapshotDB
+    public class SchemaSnapshotDB : SMOBaseClass
     {
-
-        private readonly string _connectionString;
-        private readonly SchemaSnapshotDBOptions options;
-        private readonly ScriptingOptions ScriptingOptions;
-        private string masterConnectionString
+        public SchemaSnapshotDB(DBADashConnection source, SchemaSnapshotDBOptions options) : base(source, options)
         {
-            get
-            {
-                var builder = new SqlConnectionStringBuilder(_connectionString)
-                {
-                    InitialCatalog = "master"
-                };
-                return builder.ConnectionString;
-            }
+
         }
-
-        private byte[] computeHash(byte[] obj)
+        public SchemaSnapshotDB(DBADashConnection source) : base(source)
         {
-            using(var crypt = SHA256.Create())
-            {
-                return crypt.ComputeHash(obj); 
-            }
-        }
-       
-        public SchemaSnapshotDB(string connectionString,SchemaSnapshotDBOptions options)
-        {
-            _connectionString = connectionString;
-            options = options == null ? new SchemaSnapshotDBOptions() : options;
-            this.options = options;
-            this.ScriptingOptions = options.ScriptOptions();
-        }
-
-        public SchemaSnapshotDB(string connectionString)
-        {
-            _connectionString = connectionString;
-            this.options = new SchemaSnapshotDBOptions();
-            this.ScriptingOptions = options.ScriptOptions();
-        }
-
-
-        private static string stringCollectionToString(StringCollection sc)
-        {
-            StringBuilder sb = new StringBuilder();
-            foreach (var s in sc)
-            {
-                sb.AppendLine(s);
-                sb.AppendLine("GO");
-            }
-            return sb.ToString();
-        }
-
-
-        private DataTable JobDataTableSchema()
-        {
-            DataTable dtSchema = new DataTable("Jobs");
-            dtSchema.Columns.Add("job_id", typeof(Guid));
-            dtSchema.Columns.Add("originating_server");
-            dtSchema.Columns.Add("name");
-            dtSchema.Columns.Add("enabled", typeof(bool));
-            dtSchema.Columns.Add("description");
-            dtSchema.Columns.Add("start_step_id", typeof(int));
-            dtSchema.Columns.Add("category_id", typeof(int));
-            dtSchema.Columns.Add("category");
-            dtSchema.Columns.Add("owner");
-            dtSchema.Columns.Add("notify_level_eventlog", typeof(int));
-            dtSchema.Columns.Add("notify_level_email", typeof(int));
-            dtSchema.Columns.Add("notify_level_netsend", typeof(int));
-            dtSchema.Columns.Add("notify_level_page", typeof(int));
-            dtSchema.Columns.Add("notify_email_operator");
-            dtSchema.Columns.Add("notify_netsend_operator");
-            dtSchema.Columns.Add("notify_page_operator");
-            dtSchema.Columns.Add("delete_level", typeof(int));
-            dtSchema.Columns.Add("date_created", typeof(DateTime));
-            dtSchema.Columns.Add("date_modified", typeof(DateTime));
-            dtSchema.Columns.Add("version_number", typeof(int));
-            dtSchema.Columns.Add("has_schedule", typeof(bool));
-            dtSchema.Columns.Add("has_server", typeof(bool));
-            dtSchema.Columns.Add("has_step", typeof(bool));
-            dtSchema.Columns.Add("DDLHash", typeof(byte[]));
-            dtSchema.Columns.Add("DDL", typeof(byte[]));
-            return dtSchema;
-        }
-
-        private DataTable JobStepTableSchema()
-        {
-            var jobStepDT = new DataTable("JobSteps");
-            jobStepDT.Columns.Add("job_id", typeof(Guid));
-            jobStepDT.Columns.Add("step_id", typeof(int));
-            jobStepDT.Columns.Add("step_name");
-            jobStepDT.Columns.Add("subsystem");
-            jobStepDT.Columns.Add("command");
-            jobStepDT.Columns.Add("cmdexec_success_code", typeof(int));
-            jobStepDT.Columns.Add("on_success_action", typeof(short));
-            jobStepDT.Columns.Add("on_success_step_id", typeof(int));
-            jobStepDT.Columns.Add("on_fail_action", typeof(short));
-            jobStepDT.Columns.Add("on_fail_step_id", typeof(int));
-            jobStepDT.Columns.Add("database_name");
-            jobStepDT.Columns.Add("database_user_name");
-            jobStepDT.Columns.Add("retry_attempts", typeof(int));
-            jobStepDT.Columns.Add("retry_interval", typeof(int));
-            jobStepDT.Columns.Add("output_file_name");
-            jobStepDT.Columns.Add("proxy_name");
-            return jobStepDT;
-        }
-
-        public void SnapshotJobs(ref DataSet ds)
-        {
-            var jobDT = JobDataTableSchema();
-            var jobStepDT = JobStepTableSchema();
-            List<Exception> errors = new();
-            using (var cn = new Microsoft.Data.SqlClient.SqlConnection(_connectionString))
-            {
-                var instance = new Microsoft.SqlServer.Management.Smo.Server(new Microsoft.SqlServer.Management.Common.ServerConnection(cn));
-                foreach(Microsoft.SqlServer.Management.Smo.Agent.Job job in instance.JobServer.Jobs)
-                {
-                    DataRow r = jobDT.NewRow();
-                    string sDDL;
-                    try
-                    {
-                        sDDL = stringCollectionToString(job.Script(ScriptingOptions));
-                    }
-                    catch(Exception ex)
-                    {
-                        string message = String.Format("Error scripting agent job `{0}` on {1}", job.Name, instance.Name);
-                        sDDL = "/*\n Error scripting job: \n" + ex.Message + "\n*/";
-                        errors.Add(new Exception(message, ex));
-                        Log.Error(ex, message);
-                    }
-                   
-                    var bDDL = Zip(sDDL);
-                    r["name"] = job.Name;
-                    r["job_id"] = job.JobID;
-                    r["enabled"] = job.IsEnabled;
-                    r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
-                    r["date_created"] = job.DateCreated;
-                    r["date_modified"] = job.DateLastModified;
-                    r["category"] = job.Category;
-                    r["category_id"] = job.CategoryID;
-                    r["description"] = job.Description;
-                    r["version_number"] = job.VersionNumber;
-                    r["delete_level"] = job.DeleteLevel;
-                    r["notify_level_page"] = job.DeleteLevel;
-                    r["notify_level_netsend"] = job.NetSendLevel;
-                    r["notify_level_email"] = job.EmailLevel;
-                    r["notify_level_eventlog"] = job.EventLogLevel;
-                    r["notify_email_operator"] = job.OperatorToEmail;
-                    r["notify_netsend_operator"] = job.OperatorToNetSend;
-                    r["notify_page_operator"] = job.OperatorToPage;
-                    r["owner"] = job.OwnerLoginName;
-                    r["start_step_id"] = job.StartStepID;
-                    r["has_schedule"] = job.HasSchedule;
-                    r["has_server"] = job.HasServer;
-                    r["has_step"] = job.HasStep;
-                    jobDT.Rows.Add(r);
-
-                    foreach(Microsoft.SqlServer.Management.Smo.Agent.JobStep step in job.JobSteps)
-                    {
-                        var stepR = jobStepDT.NewRow();
-                        stepR["job_id"] = job.JobID;
-                        stepR["step_name"] = step.Name;
-                        stepR["database_name"] = step.DatabaseName;
-                        stepR["step_id"] = step.ID;
-                        stepR["subsystem"] = step.SubSystem;
-                        stepR["command"] = step.Command;
-                        stepR["cmdexec_success_code"] = step.CommandExecutionSuccessCode;
-                        stepR["on_success_action"] = step.OnSuccessAction;
-                        stepR["on_success_step_id"] = step.OnSuccessStep;
-                        stepR["on_fail_action"] = step.OnFailAction;
-                        stepR["on_fail_step_id"] = step.OnFailStep;
-                        stepR["database_user_name"] = step.DatabaseUserName;
-                        stepR["retry_attempts"] = step.RetryAttempts;
-                        stepR["retry_interval"] = step.RetryInterval;
-                        stepR["output_file_name"] = step.OutputFileName;
-                        stepR["proxy_name"] = step.ProxyName;
-
-                        jobStepDT.Rows.Add(stepR);
-
-                    }
-                }
-            }
-            ds.Tables.Add(jobDT);
-            ds.Tables.Add(jobStepDT);
-            if (errors.Count > 0)
-            {
-                throw new AggregateException(errors);
-            }
+            
         }
 
         private DataTable rgDTSchema()
@@ -293,7 +113,7 @@ namespace DBADash
         {
             DataTable dtRG = rgDTSchema();
             bool reconfigError = false;
-            using (var cn = new Microsoft.Data.SqlClient.SqlConnection(masterConnectionString))
+            using (var cn = new Microsoft.Data.SqlClient.SqlConnection(MasterConnectionString))
             {
                 var instance = new Microsoft.SqlServer.Management.Smo.Server(new Microsoft.SqlServer.Management.Common.ServerConnection(cn));
                 if (instance.EngineEdition == Edition.EnterpriseOrDeveloper) //  RG is an enterprise only edition feature
@@ -302,7 +122,7 @@ namespace DBADash
                     // Script RG configuration
                     var sc = rg.Script();
                     // Add classifier function script
-                    sc.Insert(0,ObjectDDL(masterConnectionString, rg.ClassifierFunction));
+                    sc.Insert(0,ObjectDDL(MasterConnectionString, rg.ClassifierFunction));
 
                     // Script out resource pool configuration and workload groups
                     foreach (ResourcePool pool in rg.ResourcePools)
@@ -347,7 +167,7 @@ namespace DBADash
                     row["reconfiguration_error"] = reconfigError;
                     row["reconfiguration_pending"] = rg.ReconfigurePending;
                     row["max_outstanding_io_per_volume"] =rg.ServerVersion.Major >=12 ? rg.MaxOutstandingIOPerVolume : 0; // Supported 2014 and later
-                    row["script"] = stringCollectionToString(sc);
+                    row["script"] = StringCollectionToString(sc);
                     dtRG.Rows.Add(row);
                     
                 }
@@ -370,7 +190,7 @@ namespace DBADash
         public DataTable SnapshotDB(string DBName)
         {
 
-            using (var cn = new Microsoft.Data.SqlClient.SqlConnection(_connectionString))
+            using (var cn = new Microsoft.Data.SqlClient.SqlConnection(ConnectionString))
             using (var opSS = Operation.Begin("Schema snapshot {DBame}", DBName))
             {
                 DataTable dtSchema = new DataTable("Schema_" + DBName);
@@ -393,7 +213,7 @@ namespace DBADash
                 {
                     if (options.ObjectTypes.Contains(SchemaSnapshotDBObjectTypes.Database))
                     {
-                        sDDL = stringCollectionToString(db.Script(ScriptingOptions));
+                        sDDL = StringCollectionToString(db.Script(ScriptingOptions));
                         bDDL = Zip(sDDL);
                         r = dtSchema.NewRow();
                         r["ObjectName"] = "Database";
@@ -401,7 +221,7 @@ namespace DBADash
                         r["ObjectType"] = "DB";
                         r["object_id"] = 0;
                         r["DDL"] = bDDL;
-                        r["DDLHash"] = computeHash(bDDL);
+                        r["DDLHash"] = ComputeHash(bDDL);
                         r["ObjectDateCreated"] = db.CreateDate;
                         r["ObjectDateModified"] = DBNull.Value;
                         dtSchema.Rows.Add(r);
@@ -568,14 +388,14 @@ namespace DBADash
                 if (!q.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(q.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(q.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = q.Name;
                     r["SchemaName"] = q.Schema;
                     r["ObjectType"] = "SQ";
                     r["object_id"] = q.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = q.CreateDate;
                     r["ObjectDateModified"] = q.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -584,14 +404,14 @@ namespace DBADash
             foreach (ServiceRoute rt in db.ServiceBroker.Routes)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(rt.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(rt.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = rt.Name;
                 r["SchemaName"] = "";
                 r["ObjectType"] = "SBR";
                 r["object_id"] = rt.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = "1900-01-01";
                 r["ObjectDateModified"] = "1900-01-01";
                 dtSchema.Rows.Add(r);
@@ -601,14 +421,14 @@ namespace DBADash
                 if (!m.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(m.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(m.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = m.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "SBM";
                     r["object_id"] = m.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = "1900-01-01";
                     r["ObjectDateModified"] = "1900-01-01";
                     dtSchema.Rows.Add(r);
@@ -619,14 +439,14 @@ namespace DBADash
                 if (!s.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(s.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(s.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = s.Name;
                     r["SchemaName"] = s.QueueSchema;
                     r["ObjectType"] = "SBS";
                     r["object_id"] = s.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = "1900-01-01";
                     r["ObjectDateModified"] = "1900-01-01";
                     dtSchema.Rows.Add(r);
@@ -637,14 +457,14 @@ namespace DBADash
                 if (!c.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(c.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(c.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = c.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "SBC";
                     r["object_id"] = c.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = "1900-01-01";
                     r["ObjectDateModified"] = "1900-01-01";
                     dtSchema.Rows.Add(r);
@@ -653,14 +473,14 @@ namespace DBADash
             foreach (RemoteServiceBinding b in db.ServiceBroker.RemoteServiceBindings)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(b.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(b.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = b.Name;
                 r["SchemaName"] = "";
                 r["ObjectType"] = "SBB";
                 r["object_id"] = b.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = "1900-01-01";
                 r["ObjectDateModified"] = "1900-01-01";
                 dtSchema.Rows.Add(r);
@@ -670,14 +490,14 @@ namespace DBADash
                 foreach (BrokerPriority p in db.ServiceBroker.Priorities)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(p.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(p.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = p.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "SBP";
                     r["object_id"] = p.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = "1900-01-01";
                     r["ObjectDateModified"] = "1900-01-01";
                     dtSchema.Rows.Add(r);
@@ -692,14 +512,14 @@ namespace DBADash
                 foreach (Sequence s in db.Sequences)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(s.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(s.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = s.Name;
                     r["SchemaName"] = s.Schema;
                     r["ObjectType"] = "SO";
                     r["object_id"] = s.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = s.CreateDate;
                     r["ObjectDateModified"] = s.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -713,14 +533,14 @@ namespace DBADash
             foreach (ApplicationRole ar in db.ApplicationRoles)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(ar.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(ar.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = ar.Name;
                 r["SchemaName"] = "";
                 r["ObjectType"] = "ARO";
                 r["object_id"] = ar.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = ar.CreateDate;
                 r["ObjectDateModified"] = ar.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -734,14 +554,14 @@ namespace DBADash
                 if (!u.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(u.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(u.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = u.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "USR";
                     r["object_id"] = u.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = u.CreateDate;
                     r["ObjectDateModified"] = u.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -756,14 +576,14 @@ namespace DBADash
                 if (!dbr.IsFixedRole)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(dbr.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(dbr.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = dbr.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "ROL";
                     r["object_id"] = dbr.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = dbr.CreateDate;
                     r["ObjectDateModified"] = dbr.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -776,14 +596,14 @@ namespace DBADash
             foreach (UserDefinedAggregate a in db.UserDefinedAggregates)
             {  
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(a.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(a.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = a.Name;
                 r["SchemaName"] = a.Schema;
                 r["ObjectType"] = "AF";
                 r["object_id"] = a.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = a.CreateDate;
                 r["ObjectDateModified"] = a.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -797,14 +617,14 @@ namespace DBADash
                 if (!a.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(a.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(a.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = a.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "CLR";
                     r["object_id"] = a.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = a.CreateDate;
                     r["ObjectDateModified"] = DBNull.Value;
                     dtSchema.Rows.Add(r);
@@ -817,14 +637,14 @@ namespace DBADash
             foreach (XmlSchemaCollection x in db.XmlSchemaCollections)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(x.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(x.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = x.Name;
                 r["SchemaName"] = x.Schema;
                 r["ObjectType"] = "XSC";
                 r["object_id"] = x.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = x.CreateDate;
                 r["ObjectDateModified"] = x.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -838,14 +658,14 @@ namespace DBADash
                 if (!s.IsSystemObject)
                 {
                     var r = dtSchema.NewRow();
-                    var sDDL = stringCollectionToString(s.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(s.Script(ScriptingOptions));
                     var bDDL = Zip(sDDL);
                     r["ObjectName"] = s.Name;
                     r["SchemaName"] = "";
                     r["ObjectType"] = "SCH";
                     r["object_id"] = s.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = DBNull.Value;
                     r["ObjectDateModified"] = DBNull.Value;
                     dtSchema.Rows.Add(r);
@@ -859,14 +679,14 @@ namespace DBADash
             foreach (Synonym s in db.Synonyms)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(s.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(s.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = s.Name;
                 r["SchemaName"] = s.Schema;
                 r["ObjectType"] = "SN";
                 r["object_id"] = s.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = s.CreateDate;
                 r["ObjectDateModified"] = s.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -885,7 +705,7 @@ namespace DBADash
                 }
                 else
                 {
-                    sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                    sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                 }
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = t.Name;
@@ -893,7 +713,7 @@ namespace DBADash
                 r["ObjectType"] = "DTR";
                 r["object_id"] = t.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = t.CreateDate;
                 r["ObjectDateModified"] = t.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -907,14 +727,14 @@ namespace DBADash
             foreach (UserDefinedTableType t in db.UserDefinedTableTypes)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = t.Name;
                 r["SchemaName"] = t.Schema;
                 r["ObjectType"] = "TT";
                 r["object_id"] = t.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = t.CreateDate;
                 r["ObjectDateModified"] = t.DateLastModified;
                 dtSchema.Rows.Add(r);
@@ -926,14 +746,14 @@ namespace DBADash
             foreach (UserDefinedType t in db.UserDefinedTypes)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = t.Name;
                 r["SchemaName"] = t.Schema;
                 r["ObjectType"] = "UTY";
                 r["object_id"] = t.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = DBNull.Value;
                 r["ObjectDateModified"] = DBNull.Value;
                 dtSchema.Rows.Add(r);
@@ -946,14 +766,14 @@ namespace DBADash
             foreach (UserDefinedDataType t in db.UserDefinedDataTypes)
             {
                 var r = dtSchema.NewRow();
-                var sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                var sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                 var bDDL = Zip(sDDL);
                 r["ObjectName"] = t.Name;
                 r["SchemaName"] = t.Schema;
                 r["ObjectType"] = "TYP";
                 r["object_id"] = t.ID;
                 r["DDL"] = bDDL;
-                r["DDLHash"] = computeHash(bDDL);
+                r["DDLHash"] = ComputeHash(bDDL);
                 r["ObjectDateCreated"] = DBNull.Value;
                 r["ObjectDateModified"] = DBNull.Value;
                 dtSchema.Rows.Add(r);
@@ -1013,7 +833,7 @@ namespace DBADash
 
                     r["object_id"] = f.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = f.CreateDate;
                     r["ObjectDateModified"] = f.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -1036,7 +856,7 @@ namespace DBADash
                     }
                     else
                     {
-                        sDDL = stringCollectionToString(v.Script(ScriptingOptions));
+                        sDDL = StringCollectionToString(v.Script(ScriptingOptions));
                     }
                  
                     var bDDL = Zip(sDDL);
@@ -1045,7 +865,7 @@ namespace DBADash
                     r["ObjectType"] = "V";
                     r["object_id"] = v.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = v.CreateDate;
                     r["ObjectDateModified"] = v.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -1086,7 +906,7 @@ namespace DBADash
                     r["SchemaName"] = sp.Schema;
                     r["object_id"] = sp.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = sp.CreateDate;
                     r["ObjectDateModified"] = sp.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -1117,7 +937,7 @@ namespace DBADash
                             }                      
                         }                       
                     }
-                    var sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                    var sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                     if (hasEncryptedTriggers)
                     {
                         sDDL += Environment.NewLine +  "/* Encrypted Triggers */";
@@ -1128,7 +948,7 @@ namespace DBADash
                     r["ObjectType"] = "U";
                     r["object_id"] = t.ID;
                     r["DDL"] = bDDL;
-                    r["DDLHash"] = computeHash(bDDL);
+                    r["DDLHash"] = ComputeHash(bDDL);
                     r["ObjectDateCreated"] = t.CreateDate;
                     r["ObjectDateModified"] = t.DateLastModified;
                     dtSchema.Rows.Add(r);
@@ -1155,7 +975,7 @@ namespace DBADash
                         }
                         else
                         {
-                            sDDL = stringCollectionToString(t.Script(ScriptingOptions));
+                            sDDL = StringCollectionToString(t.Script(ScriptingOptions));
                         }
                         var bDDL = Zip(sDDL);
                         r["ObjectName"] = t.Name;
@@ -1163,7 +983,7 @@ namespace DBADash
                         r["ObjectType"] = t.ImplementationType == ImplementationType.SqlClr ? "TA" : "TR";
                         r["object_id"] = t.ID;
                         r["DDL"] = bDDL;
-                        r["DDLHash"] = computeHash(bDDL);
+                        r["DDLHash"] = ComputeHash(bDDL);
                         r["ObjectDateCreated"] = t.CreateDate;
                         r["ObjectDateModified"] = t.DateLastModified;
                         dtSchema.Rows.Add(r);
@@ -1173,49 +993,7 @@ namespace DBADash
         }
 
  
-        public static void CopyTo(Stream src, Stream dest)
-        {
-            byte[] bytes = new byte[4096];
 
-            int cnt;
-
-            while ((cnt = src.Read(bytes, 0, bytes.Length)) != 0)
-            {
-                dest.Write(bytes, 0, cnt);
-            }
-        }
-
-        public static byte[] Zip(string str)
-        {
-            var bytes = Encoding.Unicode.GetBytes(str);
-
-            using (var msi = new MemoryStream(bytes))
-            using (var mso = new MemoryStream())
-            {
-                using (var gs = new GZipStream(mso, CompressionMode.Compress))
-                {
-                    //msi.CopyTo(gs);
-                    CopyTo(msi, gs);
-                }
-
-                return mso.ToArray();
-            }
-        }
-
-        public static string Unzip(byte[] bytes)
-        {
-            using (var msi = new MemoryStream(bytes))
-            using (var mso = new MemoryStream())
-            {
-                using (var gs = new GZipStream(msi, CompressionMode.Decompress))
-                {
-                    //gs.CopyTo(mso);
-                    CopyTo(gs, mso);
-                }
-
-                return Encoding.Unicode.GetString(mso.ToArray());
-            }
-        }
 
     }
 }
